+++ src/Util/PaymentStatusUtil.php
@@ -177,7 +177,7 @@ class PaymentStatusUtil
         $criteria->getAssociation('transactions')->addSorting(new FieldSorting('createdAt'));
 
         /** @var OrderEntity|null $order */
-        $order = $this->orderRepository->search($criteria, $context)->first();
+        $order = $this->orderRepository->search($criteria, $context)->getEntities()->first();
 
         if ($order === null) {
             throw OrderException::orderNotFound($orderId);