+++ tests/Pos/Converter/VariantConverterTest.php
@@ -190,7 +190,7 @@ class VariantConverterTest extends TestCase
         /** @var EntityRepository $currencyRepository */
         $currencyRepository = $this->getContainer()->get('currency.repository');
         /** @var CurrencyEntity|null $currency */
-        $currency = $currencyRepository->search($criteria, Context::createDefaultContext())->first();
+        $currency = $currencyRepository->search($criteria, Context::createDefaultContext())->getEntities()->first();
 
         static::assertNotNull($currency);